Remove access without valid API key
Tellows started to block access without a valid API key (they need to feed their family as well). Currently a missing or wrong MD5 key is breaking the JSON format of the response, which led to a PHP error message.

Access without or with an invalid API key will now return a debug message, not leading to any error.