Tellows-API: Added option to respect Allowlist
In many cases an entry in the Allowlist (manually added or automatically by a dial-out) should always make it back as an inbound call without being SPAM-filtered. FreePBX has no build-in priority for Allowlist calls to bypass SPAM checks in Superfecta.

Therefore I've added a new option "Allowlist Skip" where you can choose that any numbers on the Allowlist will not be treated as SPAM (and maybe routed to a blackhole destination).